requirements.txt
streamlit
cohere
pydantic
pdfplumber
python-docx
mailparser
numpy